+++ core-primitives/substrate-sgx/sp-io/src/lib.rs
@@ -652,10 +652,8 @@ pub mod crypto {
 	) -> Result<[u8; 64], EcdsaVerifyError> {
 		let rs = libsecp256k1::Signature::parse_standard_slice(&sig[0..64])
 			.map_err(|_| EcdsaVerifyError::BadRS)?;
-		let v = libsecp256k1::RecoveryId::parse(
-			if sig[64] > 26 { sig[64] - 27 } else { sig[64] } as u8
-		)
-		.map_err(|_| EcdsaVerifyError::BadV)?;
+		let v = libsecp256k1::RecoveryId::parse(if sig[64] > 26 { sig[64] - 27 } else { sig[64] })
+			.map_err(|_| EcdsaVerifyError::BadV)?;
 		let pubkey = libsecp256k1::recover(&libsecp256k1::Message::parse(msg), &rs, &v)
 			.map_err(|_| EcdsaVerifyError::BadSignature)?;
 		let mut res = [0u8; 64];
